hotel_ms/backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy[asyncio]
asyncmy
python-multipart
python-dotenv
loguru
greenlet
passlib[bcrypt]
python-jose[cryptography]
